CCD-1717: Test build with maxParallelForks
