+++ .travis.yml
@@ -19,6 +19,7 @@ script:
 env:
   global:
     - CODACY_PROJECT_TOKEN=24f7bf62b3fa4328aa3b9bf32c07b0d9
+    - MAX_NUM_PARALLEL_THREADS=2
 
 before_install:
   - curl https://raw.githubusercontent.com/hmcts/reform-api-docs/master/bin/publish-swagger-group-docs.sh > publish-swagger-group-docs.sh
