+++ build.gradle
@@ -135,6 +135,10 @@ tasks.withType(Checkstyle) {
     exclude 'uk/gov/hmcts/ccd/domain/types/sanitiser/document/**'
 }
 
+tasks.withType(Test) {
+    maxParallelForks = System.getenv('MAX_NUM_PARALLEL_THREADS') ?: 6 as int
+}
+
 dependencies {
     implementation group: 'org.springframework.cloud', name: 'spring-cloud-starter-bootstrap', version: '3.0.1'
     compile group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign', version: '2.2.7.RELEASE'
